test FeedItemsMapper in isolation
